refactor: merge `utils-shared` into `utils`
