+++ __tests__/utils/isFunction.spec.ts
@@ -1,4 +1,4 @@
-import { isFunction } from '../src';
+import { isFunction } from '../../src/utils/index';
 
 describe('isFunction', () => {
     it('isFunction', () => {
