--- packages/utils-shared/CHANGELOG.md
@@ -1,19 +1,0 @@
-## v2.0.0
-
--   🚀 add merge for custom merge a list of objects
--   🚀 add `isPlainObject` for check value whether a plain object
--   🚀 add `isFunction` for checking value whether is a function
--   🚀 add `immutable`
--   🔧 move `getParamsByQuerystring` and `getQuerystringByParams` to `Querystring`
-
-## v1.7.2
-
--   🐛 avoid some error in querystring. Just like emoji
-
-## v1.7.1
-
--   🚀 add `isNodeJS`
--   🚀 add `getResponseStatus`
--   🚀 add `isRequestWithoutBody`
--   🚀 add `getParamsByQuerystring`
--   🚀 add `getQuerystringByParams`
