--- packages/utils-shared/README.md
@@ -1,19 +1,0 @@
-# @fatcherjs/utils-shared
-
-A shared lib for helpers functions.
-
-## Install
-
-### NPM
-
-```bash
->$ npm install @fatcherjs/utils-shared
-#or yarn
->$ yarn add @fatcherjs/utils-shared
-#or pnpm
->$ pnpm add @fatcherjs/utils-shared
-```
-
-## LICENSE
-
-[MIT](./LICENSE)
