--- packages/utils-shared/package.json
@@ -1,12 +1,0 @@
-{
-    "name": "@fatcherjs/utils-shared",
-    "version": "2.0.0",
-    "main": "dist/utils-shared.js",
-    "module": "dist/utils-shared.esm.js",
-    "typings": "dist/utils-shared.d.ts",
-    "scripts": {
-        "dev": "rimraf dist && rollup -c rollup.config.ts -w",
-        "build": "rimraf dist && rollup -c rollup.config.ts",
-        "deploy": "pnpm run build && pnpm publish --no-git-check --access public"
-    }
-}
