+++ src/request.ts
@@ -1,6 +1,6 @@
 import { RequestMethod, MiddlewareResult } from './interfaces';
 import { FatcherError } from './FatcherError';
-import { isPlainObject, Querystring } from '@fatcherjs/utils-shared';
+import { isPlainObject, Querystring } from './utils/index';
 import { defineMiddleware } from './helpers';
 
 /**
