+++ src/utils/isFunction.ts
